+++ src/MicroElements.Reflection.Sources/MicroElements/Reflection/CodeCompiler.cs
@@ -11,8 +11,11 @@ namespace MicroElements.Reflection.CodeCompiler
 {
     using System;
     using System.Collections.Concurrent;
+    using System.Collections.Generic;
+    using System.Linq;
     using System.Linq.Expressions;
     using System.Reflection;
+    using MicroElements.Reflection.FriendlyName;
 
     /// <summary>
     /// Creates compiled functions.
@@ -21,59 +24,21 @@ namespace MicroElements.Reflection.CodeCompiler
     {
         #region Caches
 
-        private readonly struct FuncKey : IEquatable<FuncKey>
-        {
-            /// <summary>
-            /// Function type.
-            /// </summary>
-            private readonly Type _type;
-
-            /// <summary>
-            /// Function name.
-            /// </summary>
-            private readonly string _name;
-
-            /// <summary>
-            /// Initializes a new instance of the <see cref="FuncKey"/> struct.
-            /// </summary>
-            /// <param name="type">Function type.</param>
-            /// <param name="name">Function name.</param>
-            public FuncKey(Type type, string name)
-            {
-                _type = type;
-                _name = name;
-            }
-
-            /// <inheritdoc />
-            public bool Equals(FuncKey other) => _name == other._name && _type == other._type;
-
-            /// <inheritdoc />
-            public override bool Equals(object? obj) => obj is FuncKey other && Equals(other);
-
-            /// <inheritdoc />
-            public override int GetHashCode() => HashCode.Combine(_name, _type);
-
-            public static bool operator ==(FuncKey left, FuncKey right) => left.Equals(right);
-
-            public static bool operator !=(FuncKey left, FuncKey right) => !left.Equals(right);
-        }
+        private readonly record struct FuncKey(Type Type, string Name);
 
         private static class Cache<Arg0, Result>
         {
-            public static readonly ConcurrentDictionary<FuncKey, Func<Arg0, Result>> FuncCache =
-                new ConcurrentDictionary<FuncKey, Func<Arg0, Result>>();
+            public static readonly ConcurrentDictionary<FuncKey, Func<Arg0, Result>> FuncCache = new ();
         }
 
         private static class Cache<Arg0, Arg1, Result>
         {
-            public static readonly ConcurrentDictionary<FuncKey, Func<Arg0, Arg1, Result>> FuncCache =
-                new ConcurrentDictionary<FuncKey, Func<Arg0, Arg1, Result>>();
+            public static readonly ConcurrentDictionary<FuncKey, Func<Arg0, Arg1, Result>> FuncCache = new ();
         }
 
         private static class Cache<Arg0, Arg1, Arg2, Result>
         {
-            public static readonly ConcurrentDictionary<FuncKey, Func<Arg0, Arg1, Arg2, Result>> FuncCache =
-                new ConcurrentDictionary<FuncKey, Func<Arg0, Arg1, Arg2, Result>>();
+            public static readonly ConcurrentDictionary<FuncKey, Func<Arg0, Arg1, Arg2, Result>> FuncCache = new ();
         }
 
         #endregion
@@ -97,7 +62,7 @@ namespace MicroElements.Reflection.CodeCompiler
         {
             if (!Cache<Arg0, Result>.FuncCache.TryGetValue(new FuncKey(type, name), out var cachedFunc))
             {
-                cachedFunc = CompileGeneric(type, genericMethodFunc);
+                cachedFunc = CompileGeneric(genericMethodFunc, type);
                 Cache<Arg0, Result>.FuncCache.TryAdd(new FuncKey(type, name), cachedFunc);
             }
 
@@ -119,7 +84,7 @@ namespace MicroElements.Reflection.CodeCompiler
         {
             if (!Cache<Arg0, Arg1, Result>.FuncCache.TryGetValue(new FuncKey(type, name), out var cachedFunc))
             {
-                cachedFunc = CompileGeneric(type, genericMethodFunc);
+                cachedFunc = CompileGeneric(genericMethodFunc, type);
                 Cache<Arg0, Arg1, Result>.FuncCache.TryAdd(new FuncKey(type, name), cachedFunc);
             }
 
@@ -142,62 +107,237 @@ namespace MicroElements.Reflection.CodeCompiler
         {
             if (!Cache<Arg0, Arg1, Arg2, Result>.FuncCache.TryGetValue(new FuncKey(type, name), out var cachedFunc))
             {
-                cachedFunc = CompileGeneric(type, genericMethodFunc);
+                cachedFunc = CompileGeneric(genericMethodFunc, type);
                 Cache<Arg0, Arg1, Arg2, Result>.FuncCache.TryAdd(new FuncKey(type, name), cachedFunc);
             }
 
             return cachedFunc;
         }
 
-        private static Func<Arg0, Result> CompileGeneric<Arg0, Result>(Type propertyType, Func<Arg0, Result> genericMethodFunc)
+        private static Func<Arg0, Result> CompileGeneric<Arg0, Result>(Func<Arg0, Result> genericMethodFunc, Type genericArg1)
         {
-            MethodInfo methodInfo = genericMethodFunc.Method.GetGenericMethodDefinition();
-            return Compile<Arg0, Result>(propertyType, methodInfo);
+            MethodInfo genericMethod = genericMethodFunc.Method.GetGenericMethodDefinition().MakeGenericMethod(genericArg1);
+            return Invoker.CompileMethod<Arg0, Result>(genericMethod);
         }
 
-        private static Func<Arg0, Arg1, Result> CompileGeneric<Arg0, Arg1, Result>(Type propertyType, Func<Arg0, Arg1, Result> genericMethodFunc)
+        private static Func<Arg0, Arg1, Result> CompileGeneric<Arg0, Arg1, Result>(Func<Arg0, Arg1, Result> genericMethodFunc, Type genericArg1)
         {
-            MethodInfo methodInfo = genericMethodFunc.Method.GetGenericMethodDefinition();
-            return Compile<Arg0, Arg1, Result>(propertyType, methodInfo);
+            MethodInfo genericMethod = genericMethodFunc.Method.GetGenericMethodDefinition().MakeGenericMethod(genericArg1);
+            return Invoker.CompileMethod<Arg0, Arg1, Result>(genericMethod);
         }
 
-        private static Func<Arg0, Arg1, Arg2, Result> CompileGeneric<Arg0, Arg1, Arg2, Result>(Type propertyType, Func<Arg0, Arg1, Arg2, Result> genericMethodFunc)
+        private static Func<Arg0, Arg1, Arg2, Result> CompileGeneric<Arg0, Arg1, Arg2, Result>(Func<Arg0, Arg1, Arg2, Result> genericMethodFunc, Type genericArg1)
         {
-            MethodInfo methodInfo = genericMethodFunc.Method.GetGenericMethodDefinition();
-            return Compile<Arg0, Arg1, Arg2, Result>(propertyType, methodInfo);
+            MethodInfo genericMethod = genericMethodFunc.Method.GetGenericMethodDefinition().MakeGenericMethod(genericArg1);
+            return Invoker.CompileMethod<Arg0, Arg1, Arg2, Result>(genericMethod);
+        }
+    }
+
+    /// <summary>
+    /// Invoker creates compiled cached delegates.
+    /// </summary>
+    internal static class Invoker
+    {
+        #region Caches
+
+        private static class FuncCache<TMethodArg1, TResult>
+        {
+            internal static ConcurrentDictionary<(Type MethodOwner, string MethodName, Type? GenericArg1, Type? GenericArg2), Func<TMethodArg1, TResult>> Cache = new();
         }
 
-        private static Func<Arg0, Result> Compile<Arg0, Result>(Type propertyType, MethodInfo methodInfo)
+        private static class FuncCache<TMethodArg1, TMethodArg2, TResult>
         {
-            MethodInfo genericMethod = methodInfo.MakeGenericMethod(propertyType);
-            ParameterExpression arg0 = Expression.Parameter(typeof(Arg0));
-            MethodCallExpression callExpression = Expression.Call(genericMethod, arg0);
+            internal static ConcurrentDictionary<(Type MethodOwner, string MethodName, Type? GenericArg1, Type? GenericArg2), Func<TMethodArg1, TMethodArg2, TResult>> Cache = new();
+        }
+
+        private static class FuncCache<TMethodArg1, TMethodArg2, TMethodArg3, TResult>
+        {
+            internal static ConcurrentDictionary<(Type MethodOwner, string MethodName, Type? GenericArg1, Type? GenericArg2), Func<TMethodArg1, TMethodArg2, TMethodArg3, TResult>> Cache = new();
+        }
+
+        #endregion
+
+        public static Func<TInstance, TMethodArg1, TResult> GetCompiledCachedMethod<TInstance, TMethodArg1, TResult>(
+            string methodName, Type? genericArg1 = null, Type? genericArg2 = null)
+        {
+            var cacheKey = (typeof(TInstance), MethodName: methodName, GenericArg1: genericArg1, GenericArg2: genericArg2);
+            return FuncCache<TInstance, TMethodArg1, TResult>.Cache.GetOrAdd(cacheKey,
+                a => CompileMethod<TInstance, TMethodArg1, TResult>(a.MethodOwner, a.MethodName, a.GenericArg1, a.GenericArg2));
+        }
+
+        public static Func<TInstance, TMethodArg1, TMethodArg2, TResult> GetCompiledCachedMethod<TInstance, TMethodArg1, TMethodArg2, TResult>(
+            string methodName, Type? genericArg1 = null, Type? genericArg2 = null)
+        {
+            var cacheKey = (typeof(TInstance), MethodName: methodName, GenericArg1: genericArg1, GenericArg2: genericArg2);
+            return FuncCache<TInstance, TMethodArg1, TMethodArg2, TResult>.Cache.GetOrAdd(cacheKey,
+                a => CompileMethod<TInstance, TMethodArg1, TMethodArg2, TResult>(a.MethodOwner, a.MethodName, a.GenericArg1, a.GenericArg2));
+        }
+
+        public static Func<TMethodArg1, TResult> GetCompiledCachedMethod<TMethodArg1, TResult>(
+            this Type methodOwner, string methodName, Type? genericArg1 = null, Type? genericArg2 = null)
+        {
+            var cacheKey = (MethodOwner: methodOwner, MethodName: methodName, GenericArg1: genericArg1, GenericArg2: genericArg2);
+            return FuncCache<TMethodArg1, TResult>.Cache.GetOrAdd(cacheKey,
+                a => CompileMethod<TMethodArg1, TResult>(a.MethodOwner, a.MethodName, a.GenericArg1, a.GenericArg2));
+        }
+
+        public static Func<TMethodArg1, TMethodArg2, TResult> GetCompiledCachedMethod<TMethodArg1, TMethodArg2, TResult>(
+            this Type methodOwner, string methodName, Type? genericArg1 = null, Type? genericArg2 = null)
+        {
+            var cacheKey = (MethodOwner: methodOwner, MethodName: methodName, GenericArg1: genericArg1, GenericArg2: genericArg2);
+            return FuncCache<TMethodArg1, TMethodArg2, TResult>.Cache.GetOrAdd(cacheKey,
+                a => CompileMethod<TMethodArg1, TMethodArg2, TResult>(a.MethodOwner, a.MethodName, a.GenericArg1, a.GenericArg2));
+        }
+
+        public static Func<TMethodArg1, TMethodArg2, TMethodArg3, TResult> GetCompiledCachedMethod<TMethodArg1, TMethodArg2, TMethodArg3, TResult>(
+            this Type methodOwner, string methodName, Type? genericArg1 = null, Type? genericArg2 = null)
+        {
+            var cacheKey = (MethodOwner: methodOwner, MethodName: methodName, GenericArg1: genericArg1, GenericArg2: genericArg2);
+            return FuncCache<TMethodArg1, TMethodArg2, TMethodArg3, TResult>.Cache.GetOrAdd(cacheKey,
+                a => CompileMethod<TMethodArg1, TMethodArg2, TMethodArg3, TResult>(a.MethodOwner, a.MethodName, a.GenericArg1, a.GenericArg2));
+        }
+
+        public static Func<TMethodArg1, TResult> GetCompiledCachedMethod<TMethodArg1, TResult>(Func<TMethodArg1, TResult> genericMethodFunc, Type? genericArg1 = null, Type? genericArg2 = null)
+        {
+            var methodInfo = FromFunc(genericMethodFunc, genericArg1, genericArg2);
+            var cacheKey = (MethodOwner: methodInfo.DeclaringType, MethodName: methodInfo.Name, GenericArg1: genericArg1, GenericArg2: genericArg2);
+            return FuncCache<TMethodArg1, TResult>.Cache.GetOrAdd(cacheKey,
+                a => CompileMethod<TMethodArg1, TResult>(a.MethodOwner, a.MethodName, a.GenericArg1, a.GenericArg2));
+        }
+
+        public static Func<TMethodArg1, TMethodArg2, TResult> GetCompiledCachedMethod<TMethodArg1, TMethodArg2, TResult>(Func<TMethodArg1, TMethodArg2, TResult> genericMethodFunc, Type? genericArg1 = null, Type? genericArg2 = null)
+        {
+            var methodInfo = FromFunc(genericMethodFunc, genericArg1, genericArg2);
+            var cacheKey = (MethodOwner: methodInfo.DeclaringType, MethodName: methodInfo.Name, GenericArg1: genericArg1, GenericArg2: genericArg2);
+            return FuncCache<TMethodArg1, TMethodArg2, TResult>.Cache.GetOrAdd(cacheKey,
+                a => CompileMethod<TMethodArg1, TMethodArg2, TResult>(a.MethodOwner, a.MethodName, a.GenericArg1, a.GenericArg2));
+        }
+
+        public static Func<TMethodArg1, TMethodArg2, TMethodArg3, TResult> GetCompiledCachedMethod<TMethodArg1, TMethodArg2, TMethodArg3, TResult>(Func<TMethodArg1, TMethodArg2, TMethodArg3, TResult> genericMethodFunc, Type? genericArg1 = null, Type? genericArg2 = null)
+        {
+            var methodInfo = FromFunc(genericMethodFunc, genericArg1, genericArg2);
+            var cacheKey = (MethodOwner: methodInfo.DeclaringType, MethodName: methodInfo.Name, GenericArg1: genericArg1, GenericArg2: genericArg2);
+            return FuncCache<TMethodArg1, TMethodArg2, TMethodArg3, TResult>.Cache.GetOrAdd(cacheKey,
+                a => CompileMethod<TMethodArg1, TMethodArg2, TMethodArg3, TResult>(a.MethodOwner, a.MethodName, a.GenericArg1, a.GenericArg2));
+        }
+
+        public static MethodInfo GetMethod(this Type methodOwner, string methodName, Type? genericArg1, Type? genericArg2)
+        {
+            BindingFlags bindingFlags = BindingFlags.Static | BindingFlags.Instance | BindingFlags.NonPublic | BindingFlags.Public;
+
+            var methodInfo = methodOwner.GetMethod(methodName, bindingFlags);
+            if (methodInfo is null)
+                throw new InvalidOperationException($"Type {methodOwner.GetFriendlyName()} has no method {methodName}");
+
+            return MakeGenericMethodIfGeneric(methodInfo, genericArg1, genericArg2);
+        }
+
+        public static MethodInfo MakeGenericMethodIfGeneric(this MethodInfo methodInfo, Type? genericArg1, Type? genericArg2)
+        {
+            if (methodInfo.IsGenericMethodDefinition)
+            {
+                Type[] methodGenericArguments = GetArgs(genericArg1, genericArg2).ToArray();
+                methodInfo = methodInfo.MakeGenericMethod(methodGenericArguments);
+            }
+
+            return methodInfo;
+
+            static IEnumerable<Type> GetArgs(Type? genericArg1, Type? genericArg2)
+            {
+                if (genericArg1 != null) yield return genericArg1;
+                if (genericArg2 != null) yield return genericArg2;
+            }
+        }
+
+        public static MethodInfo EnsureIsGenericMethodDefinition(this MethodInfo methodInfo)
+        {
+            if (!methodInfo.IsGenericMethodDefinition)
+                throw new InvalidOperationException("Method should be generic. Use method group instead of lambda.");
+            return methodInfo;
+        }
+
+        public static Func<TMethodArg1, TResult> CompileMethod<TMethodArg1, TResult>(
+            this Type methodOwner, string methodName, Type? genericArg1, Type? genericArg2)
+        {
+            var genericMethod = GetMethod(methodOwner, methodName, genericArg1, genericArg2);
+            return CompileMethod<TMethodArg1, TResult>(genericMethod);
+        }
+
+        public static Func<TMethodArg1, TMethodArg2, TResult> CompileMethod<TMethodArg1, TMethodArg2, TResult>(
+            this Type methodOwner, string methodName, Type? genericArg1, Type? genericArg2)
+        {
+            var genericMethod = GetMethod(methodOwner, methodName, genericArg1, genericArg2);
+            return CompileMethod<TMethodArg1, TMethodArg2, TResult>(genericMethod);
+        }
+
+        public static Func<TMethodArg1, TMethodArg2, TMethodArg3, TResult> CompileMethod<TMethodArg1, TMethodArg2, TMethodArg3, TResult>(
+            this Type methodOwner, string methodName, Type? genericArg1, Type? genericArg2)
+        {
+            var genericMethod = GetMethod(methodOwner, methodName, genericArg1, genericArg2);
+            return CompileMethod<TMethodArg1, TMethodArg2, TMethodArg3, TResult>(genericMethod);
+        }
+
+        public static Func<TMethodArg1, TResult> CompileMethod<TMethodArg1, TResult>(this MethodInfo method)
+        {
+            var arg1 = Expression.Parameter(typeof(TMethodArg1), "arg1");
+
+            MethodCallExpression callExpression = method.IsStatic ?
+                Expression.Call(null, method, arg1) :
+                Expression.Call(arg1, method);
+
             return Expression
-                .Lambda<Func<Arg0, Result>>(callExpression, arg0)
+                .Lambda<Func<TMethodArg1, TResult>>(callExpression, arg1)
                 .Compile();
         }
 
-        private static Func<Arg0, Arg1, Result> Compile<Arg0, Arg1, Result>(Type propertyType, MethodInfo methodInfo)
+        public static Func<TMethodArg1, TMethodArg2, TResult> CompileMethod<TMethodArg1, TMethodArg2, TResult>(this MethodInfo method)
         {
-            MethodInfo genericMethod = methodInfo.MakeGenericMethod(propertyType);
-            ParameterExpression arg0 = Expression.Parameter(typeof(Arg0));
-            ParameterExpression arg1 = Expression.Parameter(typeof(Arg1));
-            MethodCallExpression callExpression = Expression.Call(genericMethod, arg0, arg1);
+            var arg1 = Expression.Parameter(typeof(TMethodArg1), "arg1");
+            var arg2 = Expression.Parameter(typeof(TMethodArg2), "arg2");
+
+            MethodCallExpression callExpression = method.IsStatic ?
+                Expression.Call(null, method, arg1, arg2) :
+                Expression.Call(arg1, method, arg2);
+
             return Expression
-                .Lambda<Func<Arg0, Arg1, Result>>(callExpression, arg0, arg1)
+                .Lambda<Func<TMethodArg1, TMethodArg2, TResult>>(callExpression, arg1, arg2)
                 .Compile();
         }
 
-        private static Func<Arg0, Arg1, Arg2, Result> Compile<Arg0, Arg1, Arg2, Result>(Type propertyType, MethodInfo methodInfo)
+        public static Func<TMethodArg1, TMethodArg2, TMethodArg3, TResult> CompileMethod<TMethodArg1, TMethodArg2, TMethodArg3, TResult>(this MethodInfo method)
         {
-            MethodInfo genericMethod = methodInfo.MakeGenericMethod(propertyType);
-            ParameterExpression arg0 = Expression.Parameter(typeof(Arg0));
-            ParameterExpression arg1 = Expression.Parameter(typeof(Arg1));
-            ParameterExpression arg2 = Expression.Parameter(typeof(Arg2));
-            MethodCallExpression callExpression = Expression.Call(genericMethod, arg0, arg1, arg2);
+            var arg1 = Expression.Parameter(typeof(TMethodArg1), "arg1");
+            var arg2 = Expression.Parameter(typeof(TMethodArg2), "arg2");
+            var arg3 = Expression.Parameter(typeof(TMethodArg3), "arg3");
+
+            MethodCallExpression callExpression = method.IsStatic ?
+                Expression.Call(null, method, arg1, arg2, arg3) :
+                Expression.Call(arg1, method, arg2, arg3);
+
             return Expression
-                .Lambda<Func<Arg0, Arg1, Arg2, Result>>(callExpression, arg0, arg1, arg2)
+                .Lambda<Func<TMethodArg1, TMethodArg2, TMethodArg3, TResult>>(callExpression, arg1, arg2, arg3)
                 .Compile();
+        }
+
+        public static MethodInfo FromFunc<TMethodArg1, TResult>(Func<TMethodArg1, TResult> genericMethodGroup, Type? genericArg1 = null, Type? genericArg2 = null)
+        {
+            return genericMethodGroup.Method
+                .GetGenericMethodDefinition()
+                .MakeGenericMethodIfGeneric(genericArg1, genericArg2);
+        }
+
+        public static MethodInfo FromFunc<TMethodArg1, TMethodArg2, TResult>(Func<TMethodArg1, TMethodArg2, TResult> genericMethodGroup, Type? genericArg1 = null, Type? genericArg2 = null)
+        {
+            return genericMethodGroup.Method
+                .GetGenericMethodDefinition()
+                .MakeGenericMethodIfGeneric(genericArg1, genericArg2);
+        }
+
+        public static MethodInfo FromFunc<TMethodArg1, TMethodArg2, TMethodArg3, TResult>(Func<TMethodArg1, TMethodArg2, TMethodArg3, TResult> genericMethodGroup, Type? genericArg1 = null, Type? genericArg2 = null)
+        {
+            return genericMethodGroup.Method
+                .GetGenericMethodDefinition()
+                .MakeGenericMethodIfGeneric(genericArg1, genericArg2);
         }
     }
 }